python_requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-speech==2.28.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.10
langchain-core==0.3.22
langchain-ollama==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.147
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.0
ollama==0.4.4
orjson==3.10.12
packaging==24.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.36
srt==3.5.3
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
vosk==0.3.45
websockets==14.1
Werkzeug==3.1.3
yarl==1.18.3
